+++ synphot/compat.py
@@ -4,7 +4,14 @@
 import astropy
 from astropy.utils.introspection import minversion
 
-__all__ = ['ASTROPY_LT_2_0', 'ASTROPY_LT_4_0']
+try:
+    import specutils  # noqa
+except ImportError:
+    HAS_SPECUTILS = False
+else:
+    HAS_SPECUTILS = True
+
+__all__ = ['ASTROPY_LT_2_0', 'ASTROPY_LT_4_0', 'HAS_SPECUTILS']
 
 ASTROPY_LT_2_0 = not minversion(astropy, '2.0')
 ASTROPY_LT_4_0 = not minversion(astropy, '4.0')